+++ src/main/java/gov/usgs/wma/waterdata/DateTimeAccuracy.java
@@ -79,22 +79,12 @@ public enum DateTimeAccuracy {
 	}
 
 	/**
-	 * Parses a code
-	 *
-	 * @param code The text to match up against a DTA Code
-	 * @return A DateTimeAccuracy instance, which is never null.
-	 * If the text is null or no matching code is found, the UNKNOWN instance is returned.
+	 * This is the string we expect to find in the field_visit_comment to indicate this
+	 * DateTimeAccuracy
+	 * @return
 	 */
-	public static DateTimeAccuracy parseCode(final String code) {
-
-		if (code != null) {
-			for(DateTimeAccuracy acc : DateTimeAccuracy.values()) {
-				if (acc.isReal() && code.equals(acc.code)) {
-					return acc;
-				}
-			}
-		}
-
-		return UNKNOWN;
+	public String getMatchString() {
+		return matchString;
 	}
+
 }