+++ src/main/java/gov/usgs/wma/waterdata/DiscreteGroundWaterRowMapper.java
@@ -12,14 +12,12 @@ public class DiscreteGroundWaterRowMapper implements RowMapper<DiscreteGroundWat
 
 		DiscreteGroundWaterRules rules = new DiscreteGroundWaterRules();
 
-		DateTimeAccuracy dta = DateTimeAccuracy.parse(rs.getString("field_visit_comments"));
-
 		DiscreteGroundWater discreteGroundWater = new DiscreteGroundWater();
 		discreteGroundWater.setFieldVisitIdentifier(rs.getString("field_visit_identifier"));
 		discreteGroundWater.setLocationIdentifier(rs.getString("location_identifier"));
 		discreteGroundWater.setAgencyCode(rs.getString("agency_code"));
-		discreteGroundWater.setDateTimeAccuracyCode(dta.getCode());
-		discreteGroundWater.setDateTimeAccuracyText(dta.getText());
+		discreteGroundWater.setDateTimeAccuracyCode(null);  //Set by business rules
+		discreteGroundWater.setDateTimeAccuracyText(null);  //Set by business rules
 		discreteGroundWater.setStartTime(rs.getTimestamp("start_time"));
 		discreteGroundWater.setEndTime(rs.getTimestamp("end_time"));
 		discreteGroundWater.setParty(rs.getString("party"));
@@ -49,6 +47,7 @@ public class DiscreteGroundWaterRowMapper implements RowMapper<DiscreteGroundWat
 		discreteGroundWater.setGroundWaterMeasurement(rs.getString("ground_water_measurement"));
 		discreteGroundWater.setDatum(rs.getString("datum"));
 
+		//Applies all business rules
 		rules.apply(discreteGroundWater);
 
 		return discreteGroundWater;