+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/usgs/aqts-capture-discrete-loader/compare/master..aqts-capture-discrete-loader-0.1.12)
 ### Added
--   DateTimeAccuracy is now parsed out of field visit comments.
+-   DateTimeAccuracy is now parsed out of field visit comments and defaults to MINUTE.
 
 ## [0.0.1](https://github.com/usgs/aqts-capture-discrete-loader/tree/aqts-capture-discrete-loader-0.0.1)
 ### Added